checkpoint: renames "ConformanceError.ts" to "Conformance/Error.ts"
