+++ src/library/Sequence/Base64/Conformance/Error.ts
